bmcweb: Add RelatedItem to DIMM sensors
DIMM sensors that are associated with a DIMM inventory item should
report a `RelatedItem` pointing to the corresponding `Memory` resource.

Change-Id: Ia24b074027030fab493980a47f331aecf2e5e9f7
Patch Tracking Bug: b/237593984
Upstream-Status: Pending
Justification: This change is dependent on https://quanta-bmc-private-review.googlesource.com/c/meta-gsz/+/15935. It can be upstreamed if something like that is upstreamed, or it can be reworked with a better system when we have it.