Implement Inventory properties via Assembly schema on bmcweb
This commit implements Redfish Assembly schema on bmcweb.
This schema will be used to publish inventory data for FRUs which are
attached to a given Chassis and does not map to any specific schema
definition.

The properties which are published in this commit are LocationCode,
SparePartNumber, Model, SerialNumber and PartNumber.

One of the major use case to publish these properties via redfish
is for service engineers to identify the FRU and its location in the
system, which in turn will help them in repair/replacement related to
that FRU.

Tested:

root@ytbdq13-nfd01:~# wget -qO- http://localhost:18080/redfish/v1/Chassis/AgoraV2/Assembly
{
  "@odata.id": "/redfish/v1/Chassis/AgoraV2/Assembly",
  "@odata.type": "#Assembly.v1_3_0.Assembly",
  "Assemblies": [
    {
      "@odata.id": "/redfish/v1/Chassis/AgoraV2/Assembly#/Assemblies/BIOS_flash",
      "@odata.type": "#Assembly.v1_3_0.AssemblyData",
      "MemberId": "0",
      "Model": "model1",
      "Name": "BIOS_flash",
      "PartLocation": {
        "ServiceLabel": "U1000",
        "Type": "Slot"
      },
      "PartNumber": "12345a"
    },
    {
      "@odata.id": "/redfish/v1/Chassis/AgoraV2/Assembly#/Assemblies/BMC_flash",
      "@odata.type": "#Assembly.v1_3_0.AssemblyData",
      "MemberId": "1",
      "Name": "BMC_flash",
      "PartLocation": {
        "ServiceLabel": "U101",
        "Type": "Slot"
      }
    }
  ],
}

Google-Bug-Id: 235108193
review: https://gerrit.openbmc.org/c/openbmc/bmcweb/+/39574
review: https://gerrit.openbmc.org/c/openbmc/bmcweb/+/52534
Upstream-Status: Submitted
Justification: The review is in process
Signed-off-by: Gobinath Krishnamoorthy<gobinathk@google.com>
Change-Id: I611b3982c4dce6f65ac69a0fb4e3d10c3d533f05