Fix RelatedItems when fetching individual sensors
RelatedItems are missing when we get sensors without using expand. This
is because
https://github.com/openbmc/bmcweb/commit/c1d019a6056a2a0ef50e577b3139ab5a8dc49355
removes all inventory information when getting a single sensor. This fix
is going to work for RelatedItems only, since it doesn't actually read
the inventory interfaces' properties. It also defeats some of the
performance improvement in the previous commit, so an eventual fix for
this may require some more work.

Tested:
RelatedItem is present when getting single sensor:
{
  "@odata.id": "/redfish/v1/Chassis/izumi/Sensors/temperature_Core_1_CPU1",
  "@odata.type": "#Sensor.v1_2_0.Sensor",
  "Id": "Sensors_Core_1_CPU1",
  "Name": "Core 1 CPU1",
  "Reading": 46.0,
  "ReadingRangeMax": 127.0,
  "ReadingRangeMin": -128.0,
  "ReadingType": "Temperature",
  "ReadingUnits": "Cel",
  "RelatedItem": [
    {
      "@odata.id": "/redfish/v1/Systems/system/Processors/cpu1"
    }
  ],
  "Status": {
    "Health": "OK",
    "State": "Enabled"
  }
}
Sensors with $expand is unchanged.

Google-Bug-Id: 256630913
Change-Id: Iaa710008091010e0974080e0d9c22b3a4c5d0148
Signed-off-by: Shounak Mitra <shounak@google.com>

Patch Tracking Bug: b/258825174
Upstream review: https://gerrit.openbmc.org/c/openbmc/bmcweb/+/58642
Upstream-Status: Pending
Justification: This needs some work to fix in the general case and so it doesn't
undo performnace improvements upstream.