google/v1: Add setUpRedfishRoute for all handlers
Adding setUpRedfishRoute for all handlers because this method is required
to return a gRPC response. In setUpRedfishRoute, the Odata version header
is set and this is required for gRPC to return a response

Tested:

ytbjs2-n1:~# /data/milotic_tool_cli --resource_url='/google/v1/' --uid=  \
 --host host-control-node.localproductionnode.goog --port 443 \
 --auth tls_not_verify_server --command get \
 --cert_chain /root/.zatar/cert.pem  --private_key /root/.zatar/key.pem
URL is /google/v1/
Response JSON is /google/v1/{
  "@odata.id": "/google/v1",
  "@odata.type": "#GoogleServiceRoot.v1_0_0.GoogleServiceRoot",
  "Id": "Google Rest RootService",
  "Name": "Google Service Root",
  "RootOfTrustCollection": {
    "@odata.id": "/google/v1/RootOfTrustCollection"
  },
  "Version": "1.0.0"
}

Patch Tracking Bug: b/272117695
Upstream info / review:
Upstream-Status: Denied
Justification:

/google/v1 is not a redfish resource and should not even be in bmcweb.

This will remain downstream only and is in the same place as the gRPC
patch. When we create our own gRPC service, this patch will go along
with it.

Change-Id: I7f33cf4e46108cde8fb63caea5afe72a051a6c06
Signed-off-by: Willy Tu <wltu@google.com>